Create 'globus_sdk._internal', move 'lazy_import'
Create a new subpackage for internal components. As the first piece,
move the lazy importer.
